// src/app/block/ComparisonType.enum.ts
export enum ComparisonType {
  Equals = "equals",
  NotEquals = "notEquals",
  Contains = "contains",
  NotContains = "notContains",
  GreaterThan = "greaterThan",
  LessThan = "lessThan",
  StartsWith = "startsWith",
  EndsWith = "endsWith",
  LengthEquals = "lengthEquals",
  LengthGreaterThan = "lengthGreaterThan",
  LengthLessThan = "lengthLessThan",
  MatchesRegex = "matchesRegex",
  IsArray = "isArray",
  IsObject = "isObject",
  IsTruthy = "isTruthy",
  IsFalsy = "isFalsy",
  IsNumber = "isNumber",
  IsBoolean = "isBoolean",
  IsUndefined = "isUndefined",
  IsNull = "isNull",
  IsNullOrUndefined = "isNullOrUndefined",
  IsDate = "isDate",
  IsSymbol = "isSymbol",
  IsRegExp = "isRegExp",
  IsString = "isString",
}